move wrapping code to separate file
